curso_2/pyproject.toml
[project]
name = "curso-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.7.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.4",
]
